[rpc] Product.get_cases -> TestCase.filter
