+++ tcms/xmlrpc/tests/test_product.py
@@ -1,11 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from tcms.tests.factories import ProductFactory
-from tcms.tests.factories import TestCaseCategoryFactory
-from tcms.tests.factories import TestCaseFactory
-from tcms.tests.factories import TestPlanFactory
-from tcms.tests.factories import UserFactory
-from tcms.tests.factories import VersionFactory
 from tcms.xmlrpc.tests.utils import XmlrpcAPIBaseTest
 
 
@@ -30,33 +25,3 @@ class TestFilter(XmlrpcAPIBaseTest):
     def test_filter_non_existing(self):
         found = self.rpc_client.Product.filter({'name': "Non Existing"})
         self.assertEqual(0, len(found))
-
-
-class TestGetCases(XmlrpcAPIBaseTest):
-
-    def _fixture_setup(self):
-        super(TestGetCases, self)._fixture_setup()
-
-        self.tester = UserFactory(username='great tester')
-        self.product = ProductFactory(name='StarCraft')
-        self.version = VersionFactory(value='0.1', product=self.product)
-        self.plan = TestPlanFactory(name='Test product.get_cases',
-                                    owner=self.tester, author=self.tester,
-                                    product=self.product,
-                                    product_version=self.version)
-        self.case_category = TestCaseCategoryFactory(product=self.product)
-        self.cases_count = 10
-        self.cases = [TestCaseFactory(category=self.case_category, author=self.tester,
-                                      reviewer=self.tester, default_tester=None,
-                                      plan=[self.plan])
-                      for i in range(self.cases_count)]
-
-    def test_get_case_with_id(self):
-        cases = self.rpc_client.Product.get_cases(self.product.pk)
-        self.assertIsNotNone(cases)
-        self.assertEqual(len(cases), self.cases_count)
-
-    def test_get_case_with_name(self):
-        cases = self.rpc_client.Product.get_cases("StarCraft")
-        self.assertIsNotNone(cases)
-        self.assertEqual(len(cases), self.cases_count)
